+++ Changelog.md
@@ -8,6 +8,7 @@ OpenCore Changelog
 - Fixed intro animation getting stuck in OpenCanopy if an entry which returns to menu is selected before animation ends
 - Modified OpenCanopy to require presence of label images only when used due to `OC_ATTR_USE_GENERIC_LABEL_IMAGE`
 - Provided `OC_ATTR_REDUCE_MOTION` to optionally disable non-required OpenCanopy menu animations
+- Modified NVRAM logout hook to handle XML entities in string vars
 
 #### v0.9.7
 - Updated recovery_urls.txt